Add useRouteCalculation hook
